Move files and fix imports necessary to make Spyder start again
