+++ spyder/plugins/ipythonconsole/widgets/figurebrowser.py
