+++ spyder/plugins/lspmanager.py
@@ -36,7 +36,7 @@ from spyder.api.preferences import PluginConfigPage
 from spyder.widgets.helperwidgets import ItemDelegate
 from spyder.config.gui import get_font, get_color_scheme
 from spyder.utils.code_analysis.lsp_client import LSPClient
-from spyder.widgets.sourcecode.codeeditor import CodeEditor
+from spyder.plugins.editor.widgets.codeeditor import CodeEditor
 
 
 LSP_LANGUAGES = [
